Fix duplicated method names `auto-complete-processing--auto-complete-data`
Due to an error in the XML files I'm using to extract the
documentation of the methods in Processing, there were some methods
that had an incorrect name leading to duplicates appearing in
`auto-complete-processing--auto-complete-data`.

The XML files have been fixed separately in their respective repo
processing/processing-docs. See the following commit:
https://github.com/processing/processing-docs/commit/18775fc544a6c82fa4881c608d1c9a52aac73619