Use the `shr` library to parse the "<b>...</b>" found in the documentation
While this does remove the html "<b>" tags from the documentation it
also causes this package to start depending on the `shr` library.

The problem with the `shr` library is that it needs an Emacs compiled
with libxml2 and is only installed by default in Emacs 24 and I would
like to support Emacs 23 as well.